backend-e/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0

opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-distro==0.41b0


